🚧 Attempt device ref
